[enhancement] sort json decoded output
This has the (possible?) downside that if the decoded JSON is then
encoded back to the binary cfg format that output would be different
from the original cfg because the keys are in a different order. I'm not
sure if there would be any negative effect if the router imports the
keys in a different order than it exported.

Ref: https://github.com/doublerebel/nvram-cfg-parser/issues/8